Added code that implements the multithreaded part
