+++ src/main/java/nl/han/meron/oose/dea/NumberUnderTest.java
@@ -1,10 +1,10 @@
-package ???;
+package nl.han.meron.oose.dea;
 
 public class NumberUnderTest {
 
     private int number;
 
-    public int getNumber() {
+    public synchronized int getNumber() {
         return number++;
     }
 }
